@@ -11,52 +11,52 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.0",
-    "@angular/cdk": "^5.2.4",
-    "@angular/common": "5.2.0",
-    "@angular/compiler": "5.2.0",
-    "@angular/core": "5.2.0",
-    "@angular/flex-layout": "latest",
-    "@angular/forms": "5.2.0",
-    "@angular/http": "5.2.0",
-    "@angular/platform-browser": "5.2.0",
-    "@angular/platform-browser-dynamic": "5.2.0",
-    "@angular/platform-server": "5.2.0",
-    "@angular/router": "5.2.0",
-    "@ng-bootstrap/ng-bootstrap": "1.0.0-beta.9",
-    "@types/jszip": "^3.1.2",
-    "angular-typing-animation": "^0.5.4",
-    "angular2-multiselect-dropdown": "2.5.0",
+    "@angular/animations": "6.0.0",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/forms": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/platform-server": "6.0.0",
+    "@angular/router": "6.0.0",
+    "@ng-bootstrap/ng-bootstrap": "2.2.2",
+    "@types/file-saver": "^1.3.0",
+    "@types/jszip": "^3.1.3",
+    "angular2-multiselect-dropdown": "2.9.0",
     "auth0-js": "^9.3.2",
-    "bootstrap": "4.0.0",
-    "core-js": "2.4.1",
+    "bootstrap": "4.1.1",
+    "core-js": "2.5.6",
     "devicon-2.2": "^2.2.0",
-    "file-saver": "^1.3.3",
+    "file-saver": "^1.3.8",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
-    "jsoneditor": "^5.14.0",
+    "jquery": "3.3.1",
+    "jsoneditor": "^5.15.0",
     "jszip": "^3.1.5",
     "jw-bootstrap-switch-ng2": "1.0.10",
-    "ng-sidebar": "^7.0.1",
-    "ng2-parallax": "^1.0.0",
-    "ngx-chips": "1.5.9",
-    "ngx-gallery": "3.1.4",
-    "ngx-highlightjs": "^1.4.1",
-    "ngx-markdown": "^1.5.2",
-    "ngx-scrollbar": "1.5.6",
-    "nouislider": "9.2.0",
+    "ng-sidebar": "^7.1.0",
+    "ngx-chips": "1.9.2",
+    "ngx-gallery": "5.3.2",
+    "ngx-highlightjs": "^2.0.1",
+    "ngx-markdown": "^1.6.0",
+    "ngx-scrollbar": "2.0.0",
+    "nouislider": "11.1.0",
+    "path": "^0.12.7",
     "pe7-icon": "^1.0.4",
-    "popper.js": "1.12.6",
-    "prismjs": "^1.11.0",
-    "rxjs": "5.5.2",
-    "swagger-ui-dist": "^3.12.0",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "0.8.4"
+    "popper.js": "1.14.3",
+    "prismjs": "^1.14.0",
+    "rxjs": "^6.0.0",
+    "rxjs-compat": "^6.0.0",
+    "stream": "^0.0.2",
+    "swagger-ui-dist": "^3.18.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.2",
-    "@angular/compiler-cli": "5.2.0",
-    "@angular/language-service": "4.0.0",
+    "@angular/cli": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@angular/language-service": "6.0.0",
     "@types/jasmine": "2.5.45",
     "@types/node": "^6.0.101",
     "ack-angular-fx": "^2.1.1",
@@ -72,6 +72,7 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "2.4.2"
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.0"
   }
 }
