@@ -1,8 +1,8 @@
+
+import {throwError as observableThrowError,  Observable, pipe } from 'rxjs';
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpParams, HttpErrorResponse } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import { catchError } from 'rxjs/operators';
-import { map } from 'rxjs/operators';
+import { catchError ,  map } from 'rxjs/operators';
 import { environment } from '../../../../environments/environment';
 import { MicroserviceFeature, Architecture, Runtime, Schema, Capability, Tool } from './../';
 
@@ -116,7 +116,7 @@ export class MicroserviceService {
 
   private handleError(err: HttpErrorResponse | any) {
     console.error('An error occurred', err);
-    return Observable.throw(err.message || err);
+    return observableThrowError(err.message || err);
   }
 
 }
