@@ -1,6 +1,6 @@
 import { ViewChild, HostBinding, Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { ProductService, Product, AttributePipe, OptionPipe } from '../../core';
 import { ArtifactService, Artifact, Specification, SelectedValue } from '../../core/';
 import { SchemaService } from '../../core';
