@@ -1,4 +1,4 @@
-import { Observable } from 'rxjs/Rx';
+import { Observable, pipe, forkJoin } from 'rxjs';
 import { Injectable } from '@angular/core';
 import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot, ActivatedRoute, Router } from '@angular/router';
 import { MicroserviceService, UtilsService, Capability } from '../../core';
@@ -14,7 +14,7 @@ export class OptionsResolve implements Resolve<Capability[][]> {
     let supporting = this.microserviceService.getCapability('supporting');
     let infrastructure = this.microserviceService.getCapability('infrastructure');
     let processgovern = this.microserviceService.getCapability('process_governance');
-    return Observable.forkJoin([capabilities, core, supporting, infrastructure, processgovern]);
+    return forkJoin<Capability[]>([capabilities, core, supporting, infrastructure, processgovern]);
   }
 
 }
