@@ -1,12 +1,9 @@
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+import { BehaviorSubject ,  Subscription ,  Observable, of, timer } from 'rxjs';
+import { mergeMap } from 'rxjs/operators';
 import * as auth0 from 'auth0-js';
-import { Subscription } from 'rxjs/Subscription';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/observable/timer';
-import 'rxjs/add/operator/mergeMap';
+
 import { AUTH_CONFIG } from '../../../environments/environment';
 
 @Injectable()
@@ -180,14 +177,14 @@ export class AuthService {
     this.unscheduleRenewal();
     // Create and subscribe to expiration observable
     const expiresAt = JSON.parse(localStorage.getItem('expires_at'));
-    const expiresIn$ = Observable.of(expiresAt)
-      .mergeMap(
-        expires => {
-          const now = Date.now();
-          // Use timer to track delay until expiration
-          // to run the refresh at the proper time
-          return Observable.timer(Math.max(1, expires - now));
-        }
+    const expiresIn$ = of(expiresAt)
+      .pipe(
+        mergeMap(
+          expires => {
+            const now = Date.now();
+            return timer(Math.max(1, expires - now));
+          }
+        )
       );
 
     this.refreshSub = expiresIn$
