@@ -1,4 +1,4 @@
-import { trigger, state, style, transition, animate, keyframes } from '@angular/core';
+import { trigger, state, style, transition, animate, keyframes } from '@angular/animations';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 // SideNav
