@@ -1,6 +1,7 @@
+
+import {throwError as observableThrowError,  Observable, pipe } from 'rxjs';
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpParams, HttpErrorResponse } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
 import { catchError } from 'rxjs/operators';
 import { environment } from '../../../../environments/environment';
 import { Content, Plans } from './';
@@ -47,7 +48,7 @@ export class ContentService {
 
   private handleError(err: HttpErrorResponse | any) {
     console.error('An error occurred', err);
-    return Observable.throw(err.message || err);
+    return observableThrowError(err.message || err);
   }
 
 }
